requirements.txt
moralis
python-dotenv
requests
pandas
